==> notebooks/00-use-sherlock-out-of-the-box.ipynb <==
# Using Sherlock out-of-the-box
This notebook shows how to predict a semantic type for a given table column.
The steps are basically:
- Download files for word embedding and paragraph vector feature extraction (downloads only once) and initialize feature extraction models.
- Extract features from table columns.
- Initialize Sherlock.
- Make a prediction for the feature representation of the column.

import numpy as np
import pandas as pd
import pyarrow as pa

from sherlock import helpers
from sherlock.deploy.model import SherlockModel
from sherlock.functional import extract_features_to_csv
from sherlock.features.paragraph_vectors import initialise_pretrained_model, initialise_nltk
from sherlock.features.preprocessing import (
    extract_features,
    convert_string_lists_to_lists,
    prepare_feature_extraction,
    load_parquet_values,
)
from sherlock.features.word_embeddings import initialise_word_embeddings

#%env PYTHONHASHSEED

from pprint import pprint
from sherlock.deploy import helpers

feature_cols = helpers.categorize_features()
pprint(feature_cols['word']) # char word par rest



## Initialize feature extraction models

initialise_pretrained_model(400)

prepare_feature_extraction()
initialise_word_embeddings()
initialise_pretrained_model(400)
initialise_nltk()

## Extract features

data = pd.Series(
    [
        ["Jane Smith", "Lute Ahorn", "Anna James"],
        ["Amsterdam", "Haarlem", "Zwolle"],
        ["Chabot Street 19", "1200 fifth Avenue", "Binnenkant 22, 1011BH"],
        ["1","2","3",],
    ],
    name="values"
)

data.shape

data = pd.Series(
    [
        ["123213", "15457", "563"],
        ["Amsterdam", "Haarlem", "Zwolle"],
        ["Chabot Street 19", "1200 fifth Avenue", "Binnenkant 22, 1011BH"],
        ['0.0', '1.0', '2.0', '3.0']

    ],
    name="values"
)

extract_features(
    "../temporary.csv",
    data
)
feature_vectors = pd.read_csv("../temporary.csv", dtype=np.float32)

feature_vectors

## Initialize Sherlock

model = SherlockModel();
model.initialize_model_from_json(with_weights=True, model_id="sherlock");

## Predict semantic type for column

predicted_labels = model.predict(feature_vectors, "sherlock")

predicted_labels
