==> notebooks/03_Doc2Vec Sample Code.ipynb <==
#https://medium.com/red-buffer/doc2vec-computing-similarity-between-the-documents-47daf6c828cd
import gensim
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from nltk.tokenize import word_tokenize
from gensim.models.doc2vec import Doc2Vec

# Training data w/ 3 paragraph

data = ["The process of searching for a job can be very stressful, but it doesn’t have to be. Start with a\
        well-written resume that has appropriate keywords for your occupation. Next, conduct a targeted job search\
        for positions that meet your needs.",
        
        
        "Gardening in mixed beds is a great way to get the most productivity from a small space. Some investment\
        is required, to purchase materials for the beds themselves, as well as soil and compost. The\
        investment will likely pay-off in terms of increased productivity.",
        
        
        "Looking for a job can be very stressful, but it doesn’t have to be. Begin by writing a good resume with\
        appropriate keywords for your occupation. Second, target your job search for positions that match your\
        needs."]

tagged_data = [TaggedDocument(words=word_tokenize(_d.lower()), tags=[str(i)]) for i, _d in enumerate(data)]

print (tagged_data)
#tagged_data[0]
# tagged_data[0].words
# #tagged_data[0].tags

model = gensim.models.doc2vec.Doc2Vec(vector_size=30, min_count=2, epochs=80)
model.build_vocab(tagged_data)
model.train(tagged_data, total_examples=model.corpus_count, epochs=80)
model.save("d2v.model")

# Finding the most similar sentence with the sentence[0]
similar_doc = model.docvecs.most_similar('0')

print(similar_doc)

print(tagged_data[0].words)

print(tagged_data[2].words)


# Model Concat - #https://towardsdatascience.com/multi-class-text-classification-with-doc2vec-logistic-regression-9da9947b43f4#:~:text=Doc2vec%20is%20an%20NLP%20tool,advisable%20to%20understand%20word2vec%20approach.
