==> notebooks/05-1-use-pretrained-model.ipynb <==
import jax
from typing import Any, Callable, Sequence, Optional
from jax import lax, random, numpy as jnp
import flax
from flax import linen as nn
import optax
import numpy as np
import pandas as pd
from sklearn.metrics import f1_score, classification_report, accuracy_score
from sherlock.deploy import helpers

class RestModel(nn.Module):
    features: Sequence[int]
  
    @nn.compact
    def __call__(self, x):
        
        x = nn.BatchNorm(use_running_average=True,
                 momentum=0.9,
                 epsilon=1e-5,
                 dtype=jnp.float32)(x)
        
        return x
    
class SubModel(nn.Module):
    features: Sequence[int]
    training: bool = True
    
    @nn.compact
    def __call__(self, x):
        
        # batchnormalisation - https://github.com/google/flax/issues/932
        x = nn.BatchNorm(use_running_average=True,
                 momentum=0.9,
                 epsilon=1e-5,
                 dtype=jnp.float32)(x)
        
        x = nn.relu(nn.Dense(self.features[0])(x))
        
        # dropout
        x = nn.Dropout(rate=0.35)(x, deterministic=True)
                
        x = nn.relu(nn.Dense(self.features[1])(x)) 
        
        # todo: add 
        
        
        return x
    
class MainModel(nn.Module):
    feature_size: int = 500
    num_classes: int = 78
        
    @nn.compact
    def __call__(self, x1, x2, x3, x4):
       
        # [1] define shape        
        y1 = SubModel([300, 300], name='char_model')(x1)      
        y2 = SubModel([200, 200], name='word_model')(x2)
        y3 = SubModel([400, 400], name='par_model')(x3)
        y4 = RestModel([27], name='rest_model')(x4)
                      
        # [2] concat submodels    
        x = jnp.concatenate((y1, y2, y3, y4), axis=-1)
        
        print("check mainmodel shape")
        print(np.shape(x))
        
        # batchnormalisation
        x = nn.BatchNorm(use_running_average=True,
                 momentum=0.9,
                 epsilon=1e-5,
                 dtype=jnp.float32)(x)
        
        # dense 1
        x = nn.relu(nn.Dense(self.feature_size)(x))
        
        # dropout
        x = nn.Dropout(rate=0.35)(x, deterministic=True)
        
        # dense 2
        x = nn.relu(nn.Dense(self.feature_size)(x))
        
        # dense w/ softmax - todo: check
        x = nn.softmax(nn.Dense(self.feature_size)(x), axis=-1)
        
        return nn.Dense(self.num_classes)(x)


from flax.training import train_state 

learning_rate = 0.0001
n_training_steps = 100

def create_train_state(rng):
  mainmodel = MainModel()
  params = mainmodel.init(jax.random.PRNGKey(0), jnp.ones((1, 960)), jnp.ones((1, 201)), jnp.ones((1, 400)), jnp.ones((1, 27)))
  tx=optax.adam(learning_rate=learning_rate)
  state = train_state.TrainState.create(apply_fn=mainmodel.apply,
                                        params=params,
                                        tx=tx)
  return state 



## RESTORE
# file - byte -> TrainState

rng = jax.random.PRNGKey(0)
rng, init_rng = jax.random.split(rng)

with open("flax_model_MAR.msgpack", mode='rb') as file:
    read_data = file.read()

new_state = create_train_state(init_rng)    
restored_state = flax.serialization.from_bytes(new_state, read_data)  ##TrainState

## Prediction

# todo raw data -> featurization

feature_cols = helpers.categorize_features()

X_test = pd.read_parquet('../data/data/processed/inputX.parquet')
# y_validation = pd.read_parquet('../data/data/raw/inputY_pq.parquet').values.flatten()
y_val = pd.read_parquet('../data/data/raw/inputY_pq.parquet')
y_val = y_val.values.flatten() #to ndarray


# y_val = y_val.applymap(str)
y_validation = np.array([x.lower() for x in y_val])



X_char = X_test[feature_cols["char"]]
X_word = X_test[feature_cols["word"]]
X_par = X_test[feature_cols["par"]]
X_rest = X_test[feature_cols["rest"]]


j_v_1 = jnp.array(pd.DataFrame(X_char).to_numpy())
j_v_2 = jnp.array(pd.DataFrame(X_word).to_numpy())
j_v_3 = jnp.array(pd.DataFrame(X_par).to_numpy())
j_v_4 = jnp.array(pd.DataFrame(X_rest).to_numpy())

Prediction

mainmodel = MainModel()
y_pred_sv = mainmodel.apply(new_state.params, j_v_1, j_v_2, j_v_3, j_v_4)

y_pred_classes = helpers._proba_to_classes(y_pred_sv, "sherlock")

y_pred_classes

Score - additional

print(f1_score(y_validation, y_pred_classes, average="weighted"))
print(accuracy_score(y_validation, y_pred_classes))

print(y_validation)
print(y_pred_classes)
